+++ nym-vpn-core/crates/nym-vpn-store/src/keys/wireguard/persistence/ephemeral.rs
@@ -57,4 +57,9 @@ impl WireguardKeyStore for InMemEphemeralKeys {
             Ok(keys)
         }
     }
+
+    async fn clear_keys(&self) -> Result<(), Self::StorageError> {
+        self.keys.lock().await.clear();
+        Ok(())
+    }
 }