+++ nym-vpn-core/crates/nym-vpn-store/src/keys/wireguard/persistence/on_disk.rs
@@ -75,6 +75,11 @@ impl WireguardKeyStore for OnDiskKeys {
         };
         Ok(keys)
     }
+
+    async fn clear_keys(&self) -> Result<(), Self::StorageError> {
+        self.delete_keys().await?;
+        Ok(())
+    }
 }
 
 // all SQL goes here
@@ -146,6 +151,17 @@ impl OnDiskKeys {
             keys.exit_private_key_bs58,
             keys.expiration_time,
             keys.gateway_id_bs58,
+        )
+        .execute(&self.connection_pool)
+        .await?;
+        Ok(())
+    }
+
+    pub(crate) async fn delete_keys(&self) -> Result<(), sqlx::Error> {
+        sqlx::query!(
+            r#"
+                DELETE FROM wireguard_gateway_keys;
+            "#,
         )
         .execute(&self.connection_pool)
         .await?;